refactor(structure) restructure project as per new system design - clear separation of api, core domains, data models, infra - not yet refactored the functionality
