+++ .gitignore
@@ -332,3 +332,27 @@ internal_docs/
 
 # .ruff cache
 .ruff_cache/
+
+# Exclude private SSH keys
+id_rsa
+id_rsa.pub
+
+# Exclude additional environment files
+.env.local
+.env.development
+.env.production
+.env.staging
+.env.*.local
+
+# Exclude editor-specific files
+.vscode/
+!.vscode/settings.json  # Keep settings.json if needed
+
+# Exclude OS-specific files
+.DS_Store
+Thumbs.db
+
+# Exclude npm and Yarn debug log files
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
