--- src/api/v0/content/crawl_router.py
@@ -1,2 +1,0 @@
-# Document loading routes go here
-# Add, list, delete, etc.
