--- src/api/v0/content/document_router.py
@@ -1,1 +1,0 @@
-# Endpoints for document management go here
