+++ src/core/_exceptions.py
