+++ src/core/chat/tool_definitions.py
