+++ src/infrastructure/common/output_formatter.py
